+++ example/src/MutableReference/Prop.hs
@@ -128,10 +128,10 @@ prop_shrinkParallelMinimal = shrinkPropertyHelper (prop_referencesParallel RaceC
   hasMinimalShrink :: Fork (Program Action) -> Bool
   hasMinimalShrink
     = anyTree isMinimal
-    . unfoldTree (id &&& shrinker')
+    . unfoldTree (id &&& shrinker'')
     where
-    shrinker' :: Fork (Program Action) -> [Fork (Program Action)]
-    shrinker'
+    shrinker'' :: Fork (Program Action) -> [Fork (Program Action)]
+    shrinker''
       = map unParallelProgram
       . shrinkParallelProgram shrinker precondition transition initModel
       . ParallelProgram
@@ -184,7 +184,7 @@ instance Read (Internal Action) where
 
   readListPrec = readListPrecDefault
 
-deriving instance Eq1 v => Eq   (Action v resp)
+deriving instance Eq1 v => Eq (Action v resp)
 
 instance Eq (Internal Action) where
   Internal act1 sym1 == Internal act2 sym2 =