+++ example/test/TicketDispenserSpec.hs
@@ -1,7 +1,7 @@
 module TicketDispenserSpec (spec) where
 
 import           Test.Hspec
-                   (Spec, describe, it, xit)
+                   (Spec, describe, it)
 import           Test.Hspec.QuickCheck
                    (modifyMaxSuccess)
 
@@ -18,9 +18,12 @@ spec = do
 
   describe "Parallel property" $ do
 
-    modifyMaxSuccess (const 10) $
+    modifyMaxSuccess (const 5) $
+
       it "`prop_ticketDispenserParallelOK`: works with exclusive file locks"
         prop_ticketDispenserParallelOK
 
-    xit "`prop_ticketDispenserParallelBad`: counterexample is found when file locks are shared"
-      prop_ticketDispenserParallelBad
+    modifyMaxSuccess (const 5) $
+
+      it "`prop_ticketDispenserParallelBad`: counterexample is found when file locks are shared"
+        prop_ticketDispenserParallelBad